BFT init
protope almost ready in main.c

Screenmanager loads a logo screen and a backdrop before drawing in a platformer game.
The logo needs to be added.
Title screen added
Platformer needs the camera set to follow the player
Gameplay needs to be implemented and testing for level completion
May shift to being a 4 level Coin gathering platformer with story elements to convey the intention of the final game BFT-Z